preserveDescribeNesting; namedFiles; output polish
- config.preserveDescribeNesting:true results in all nested `describe`
sections displayed with increased indentation level just as in the
original jasmine runner (and in the specs themselves) instead of
displaying each `describe` section separately and with the same
indentation. Top-level descriptions are emphasised in this mode.

- config.namedFiles:true results in files being named and put right
into outputDir instead of a subdir with index.html

- config.reportName is the same as config.middlePathDir

- general polishing of the output: report name is written to page
title, added report date, updated some text towards jasmine2.0 (like
pending instead of skipped)